refactor(breadcrumb): use functions instead of classes and rename files
