+++ src/components/Breadcrumbs/BreadcrumbModel.ts
@@ -1,0 +1,33 @@
+/**
+ * SubjectType enum
+ * @enum {string}
+ */
+export enum SubjectType {
+  ELA = 'ELA',
+  MATH = 'MATH'
+}
+
+/**
+ * GradeType enum
+ * @enum {string}
+ */
+export enum GradeType {
+  Three = '3',
+  Four = '4',
+  Five = '5',
+  Six = '6',
+  Seven = '7',
+  Eight = '8',
+  HighSchool = 'High School'
+}
+
+/**
+ * ClaimType
+ * @enum {string}
+ */
+export enum ClaimType {
+  One = '1',
+  Two = '2',
+  Three = '3',
+  Four = '4'
+}
